+++ src/matriz.c
@@ -18,12 +18,12 @@
  * Atributos:
  * int lin: numero de linhas da matriz
  * int col: numero de colunas da matriz
- * float* dados: ponteiro para os dados da matriz, armazenados em formato de vetor
+ * double* dados: ponteiro para os dados da matriz, armazenados em formato de vetor
  */
 struct matriz_t {
     int lin;
     int col;
-    float* dados; /* Implementamos como vetor unidimensional alocado dinamicamente */
+    double* dados; /* Implementamos como vetor unidimensional alocado dinamicamente */
 };
 
 Matriz* matriz_cria(int m, int n) {
@@ -36,7 +36,7 @@ Matriz* matriz_cria(int m, int n) {
 	mat->lin = m;
 	mat->col = n;
 
-	mat->dados = (float*)malloc(m * n * sizeof(float));
+	mat->dados = (double*)malloc(m * n * sizeof(double));
 	if (mat->dados == NULL) {
 		printf("Memoria insuficiente!\n");
 		free(mat); /* Libera espaco previamente alocado para a matriz */
@@ -64,7 +64,7 @@ int matriz_colunas(Matriz* mat) {
 	return mat->col;
 }
 
-float matriz_acessa(Matriz* mat, int i, int j) {
+double matriz_acessa(Matriz* mat, int i, int j) {
 	if (i < 0 || i >= mat->lin || j < 0 || j >= mat->col) {
 		printf("Indice fora dos limites da matriz!\n");
 		exit(1);
@@ -72,7 +72,7 @@ float matriz_acessa(Matriz* mat, int i, int j) {
 	return mat->dados[i * mat->col + j]; /* Acessa o elemento usando row-major indexing */
 }
 
-void matriz_atribui(Matriz* mat, int i, int j, float valor) {
+void matriz_atribui(Matriz* mat, int i, int j, double valor) {
 	if (i < 0 || i >= mat->lin || j < 0 || j >= mat->col) {
 		printf("Indice fora dos limites da matriz!\n");
 		exit(1);
